+++ controllers/gateway/controller.go
@@ -625,7 +625,8 @@ func setDataPlaneOptionsDefaults(opts *operatorv1beta1.DataPlaneOptions, default
 		})
 	}
 
-	if opts.Deployment.Replicas == nil {
+	// If no replicas are set, set it to default 1, but only if Scaling is not set as well.
+	if opts.Deployment.Replicas == nil && opts.Deployment.Scaling == nil {
 		opts.Deployment.Replicas = lo.ToPtr(int32(1))
 	}
 }